🎨 design(#168, #41): 공유 대시보드 비로그인 상태일때 레이아웃 꽉 차게
